server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
openai==1.44.0
google-generativeai==0.7.2
httpx==0.27.2
pydantic-settings==2.4.0
python-multipart==0.0.9
